src/requirements.txt
flask
opencv-python
torch
transformers
Pillow
